Fix registry and other bugs
